Bring in snapshot of the MdePkg from Tianocore's EDK2 project at git hash 7babb4372e6a34cbbc54249b25056272a5a9924c From 2017-Mar-03.
EDK2 is Intel's BSD Licensed UEFI implementation. We'll be bringing in
various routines from there rather than reimplementing them from
scratch for libefivar and the EFI boot loader.

The upstream repo has ^M ending on everything (sometimes multiple
times!), so the following script was run prior to import so changes we
have to do don't first include changing every line:

% find . -type f | xargs -n 1 sed -I.BAK -e `printf "s/\r//g"`
% find . -name \*.BAK | xargs rm

Also, only the MdePkg was brought in (it's 17MB, while the entire repo
is 250MB). It's almost completely certain nothing else will be used,
but if it is, it can be brough in along side MdePkg in the future.

Obtained from: https://github.com/tianocore/edk2.git


git-svn-id: https://svn.freebsd.org/base/vendor/edk2/MdePkg@314879 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f