Finish implementing -d/--device/--device-path flag to print variable as if it were a device path.
Remove language about a=b syntax on the command line. This will not be
implemented due to its limited usefulness. UEFI variables are binary
blobs, on the whole, and a simple work around exists for
strings. Clarify that the new value of the variable is taken from
stdin. Update manual with history.

Sponsored by: Netflix


git-svn-id: https://svn.freebsd.org/base/head@314927 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f