MFC r314810:
pf: Fix a crash in low-memory situations

If the call to pf_state_key_clone() in pf_get_translation() fails (i.e. there's
no more memory for it) it frees skp. This is wrong, because skp is a
pf_state_key **, so we need to free *skp, as is done later in the function.
Getting it wrong means we try to free a stack variable of the calling
pf_test_rule() function, and we panic.


git-svn-id: https://svn.freebsd.org/base/stable/10@314940 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f