qlxgbe: add GCC_MS_EXTENSIONS to CFLAGS to make old base GCC happy
The module uses unnamed structure and union fields and base GCC in
stable/10 doesn't like it.
I think that that is a C11 feature, so it is courteous of more modern
compilers to not complain about it when compiling in C99 mode.

This is a direct commit as the change is not required in head.

Approved by:	davidcs


git-svn-id: https://svn.freebsd.org/base/stable/10@315069 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f