MFC r314233:
Parameterize out the length of struct filed->f_lasttime as `MAXDATELEN`

This removes the hardcoded value for the field (16) and the equivalent
hardcoded lengths in logmsg(..).

This change is being done to help stage future work to add RFC5424/RFC5434
support to syslogd(8).

Obtained from:	Isilon OneFS (dcd33d13da) (as part of a larger change)


git-svn-id: https://svn.freebsd.org/base/stable/10@315124 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f