MFC r313683:
lockmgr: implement fast path

The main lockmgr routine takes 8 arguments which makes it impossible to
tail-call it by the intermediate vop_stdlock/unlock routines.

The routine itself starts with an if-forest and reads from the lock itself
several times.

This slows things down both single- and multi-threaded. With the patch
single-threaded fstats go 4% up and multithreaded up to ~27%.

Note that there is still a lot of room for improvement.


git-svn-id: https://svn.freebsd.org/base/stable/11@315375 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f