MFC r313853,r313859:
locks: remove SCHEDULER_STOPPED checks from primitives for modules

They all fallback to the slow path if necessary and the check is there.

This means a panicked kernel executing code from modules will be able to
succeed doing actual lock/unlock, but this was already the case for core code
which has said primitives inlined.

==

Introduce SCHEDULER_STOPPED_TD for use when the thread pointer was already read

Sprinkle in few places.


git-svn-id: https://svn.freebsd.org/base/stable/11@315386 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f