MFC r311284:
Fix bug that would result in a kernel crash in some cases involving
a symlink and an autofs mount request.  The crash was caused by namei()
calling bcopy() with a negative length, caused by numeric underflow:
in lookup(), in the relookup path, the ni_pathlen was decremented too
many times.  The bug was introduced in r296715.

Big thanks to Alex Deiter for his help with debugging this.


git-svn-id: https://svn.freebsd.org/base/stable/11@315519 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f