MFC r313350:
In r290196 the root mount hold mechanism was changed to make it not wait
for mount hold release if the root device already exists.  So, unless your
rootdev is not on USB - ie in the usual case - the root mount won't wait
for USB.  However, the old behaviour was sometimes used as "wait until USB
is fully enumerated", and r290196 broke that.

This commit adds vfs.root_mount_always_wait tunable, to force the kernel
to always wait for root mount holds, even if the root is already there.

Relnotes:	yes


git-svn-id: https://svn.freebsd.org/base/stable/11@315539 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f