MFC r313351:
Make root_mount_hold() work after boot. This is important for two
reasons. First is rerooting into USB-mounted device that happens
to be not yet enumerated. The second is when mounting with (non-root)
filesystem on USB device on a hub that's enumerated later than the root
mount: the rc scripts explicitly mount for the root mount holds to be
released, but each USB bus takes the hold asynchronously, and if that
happens after root mount, it would just get ignored.


git-svn-id: https://svn.freebsd.org/base/stable/11@315540 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f